requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==3.2
firebase-admin==6.5.0
gcloud==0.18.3
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.8
jwcrypto==1.5.6
msgpack==1.0.8
oauth2client==4.1.3
polyline==2.0.2
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyparsing==3.1.4
Pyrebase4==4.8.0
python-jwt==4.1.0
pytz==2024.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20